Version 1.0.011.0
Some updates of user mode components.

* MountTool\ArsenalImageMounterMountTool.exe is now a self-containing
exe that contains everything needed to run the MountTool GUI. It even
contains DiscUtils.dll built-in and also installs the driver components
if not already installed. It still requires the correct version of
libewf.dll in same directory to be able to mount ewf images, though.

* MountTool\ArsenalImageMounterControl.exe console tool is now somewhat
simpler to use, since it is now self-contained and does not require the
dll file to be present in same directory as the exe file.

Also added a self-contained driver setup exe and a simple zip file with
just files necessary when integrating driver setup in another driver
setup package.

// Olof Lagerkvist